chore: move fixture to a separate file
